requirements.txt
Django==2.0.5
psycopg2-binary
django-choice-object
django-hamlpy
django-bootstrap3
